added styling and updated delete post function

+++ app/posts/delete.php
@@ -10,11 +10,25 @@ if (isset($_POST['delete'])) {
     $postId = filter_var($_POST['delete'], FILTER_SANITIZE_NUMBER_INT);
     $post = getPostData($pdo, $postId);
 
-    $statement = $pdo->prepare('DELETE FROM posts WHERE id = :id AND user_id = :user_id');
-    $statement->execute([
-        ':id' => $postId,
-        ':user_id' => $_SESSION['user']['id']
-    ]);
+    if ($post["user_id"] !== $_SESSION['user']['id']) {
+        $_SESSION['errors'][] = 'You can\'t delete other users posts';
+    }
 
-    redirect('../../profile.php?id=' . $_SESSION['user']['id']);
+    if (!isset($_SESSION['errors'])) {
+        // die(var_dump($post['post_image']));
+        unlink('../..' . $post['post_image']);
+
+        $statement = $pdo->prepare('DELETE FROM posts WHERE id = :id AND user_id = :user_id');
+        $statement->execute([
+            ':id' => $postId,
+            ':user_id' => $_SESSION['user']['id']
+        ]);
+
+        $statement = $pdo->prepare('DELETE FROM roses WHERE post_id = :post_id');
+        $statement->execute([
+            ':post_id' => $postId,
+        ]);
+        redirect('../../profile.php?username=' . $_SESSION['user']['username']);
+    }
+    redirect('../../post.php?id=' . $postId);
 }
